Adding comments to new classes and methods added as part of this feature Removing SegmentMetadataFetcher as it seemed redundant Refactoring code to save failed segment reload status API calls as part of response
